Move internal functions to internals.rs
